Neatening into setup.js
