+++ src/client/views/index.pug
@@ -70,8 +70,7 @@ html(lang='en')
             a(href="https://github.com/goatonabicycle/note-generation", target="_blank") Code here
             br                       
 
-            script(src="js/options.js")
             script(src="js/patterns.js")
-            script(src="js/buttons.js")
+            script(src="js/setup.js", type="module")
